requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-core
langchain-pinecone
openai>=1.10.0,<2.0.0
pinecone-client
pypdf
pydantic>=2.0.0
